Merge protobuf support into main branch

+++ src/sancov.rs
@@ -377,8 +377,8 @@ macro_rules! handle_cmp {
         }
 
         if sizeof_type >= 8
-            && $arg1.inner() <= u32::MAX.try_into().unwrap()
-            && $arg2.inner() <= u32::MAX.try_into().unwrap()
+            && $arg1.inner() <= u64::MAX.try_into().unwrap()
+            && $arg2.inner() <= u64::MAX.try_into().unwrap()
         {
             constants.u64cov.insert((
                 $arg1.try_into().expect("failed to convert to cmp args u64"),
